pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scour"
version = "0.1.0"
description = "TUI tool for competitive research and inspiration lookup"
requires-python = ">=3.11"
dependencies = [
    "textual>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
scour = "scour.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/scour"]
